main/requirements.txt
discord.py
google-cloud-compute
google-cloud-storage
google-cloud-logging
paramiko
python-a2s